break engines/ lock effects now use game config

+++ Assets/Board Game Assets/Scripts/Board Game/GameConfig/EnginesConfig.cs
@@ -1,8 +1,8 @@
 using Newtonsoft.Json;
 
 public class EnginesConfig : ConfigData {
-    public float turnsToBreakPlayerEngine = 1f;
-    public float turnsToBreakOthersEngine = 1f;
+    public int turnsToBreakPlayerEngine = 1;
+    public int turnsToBreakOthersEngine = 1;
 
     public static EnginesConfig FromJson(string json) {
         return JsonConvert.DeserializeObject<EnginesConfig>(json);
